requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
backports.asyncio.runner==1.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
dnspython==2.7.0
dotenv==0.9.9
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
